Address #5552 review findings: staged CDX writes, record-number range check, input validation
Fixes eight findings from PR #5552's review round (chatgpt-codex-connector
and copilot-pull-request-reviewer, several reported independently by
both):

- The DBF has_production_index flag was being set before the CDX file
  was known to be durably written, so a write failure left the table
  falsely claiming a structural index it didn't have. The writer now
  stages the CDX content in a temporary sibling file, renames it over
  the real destination only once every byte is confirmed written, and
  sets the DBF flag only after that succeeds -- also fixing the
  related non-atomic-overwrite finding (opening the destination
  directly with ios::trunc could destroy an existing usable CDX before
  the new content was confirmed durable).
- CdxIndexEntry::record_number was already std::uint8_t, so a caller
  converting a record number of 256 before constructing the struct got
  a silently wrapped 0 -- the documented "fails closed above 255"
  behavior was unenforceable because the out-of-range information was
  already gone by the time this function saw it. The field is now a
  wider std::uint32_t, validated explicitly before narrowing.
- Tag names of 481-512 bytes passed the old length check but then
  overwrote the tag-table page's own required header fields once
  right-aligned to the page end; the writer now caps tag names at 480
  bytes.
- Key expressions of 512+ bytes were silently truncated (and exactly
  512 bytes left no NUL terminator) instead of failing closed.
- key_length was never validated; 0 or a value larger than one page
  produces a file this codebase's own reader can't parse at all.

All five fixes are covered by new regression tests in
tests/test_vfp_assets.cpp, including one that forces the CDX write
itself to fail and confirms the DBF flag stays unset. New locale keys
added to all four locale files; docs/77 records this as a "PR review
follow-up" section.

Verified: cmake --build build -j4 (clean, all targets), ./build/tests/
test_vfp_assets and ./build/tests/test_dbf_table directly, and a full
ctest run (100% passed, 394/394, matching the prior clean baseline).

+++ include/copperfin/vfp/cdx_writer.h
@@ -49,9 +49,15 @@ namespace copperfin::vfp {
 // oversight -- fails closed with a structured error rather than
 // producing a file whose correctness for that case was never verified.
 struct CdxIndexEntry {
-    // 0-255 only, this slice's scope (see the "record numbers beyond
-    // 255" gap in docs/77).
-    std::uint8_t record_number = 0;
+    // Deliberately wider than the single byte the leaf entry format can
+    // actually hold (0-255, see the "record numbers beyond 255" gap in
+    // docs/77): a caller-side narrowing straight to std::uint8_t would
+    // silently wrap a record number of, say, 256 down to 0 before this
+    // API could ever see the out-of-range value. create_vfp_cdx_single_
+    // tag_index_file() itself validates this and fails closed
+    // (Vfp.CdxWriter.Error.RecordNumberOutOfRange) rather than narrowing
+    // an unchecked value.
+    std::uint32_t record_number = 0;
     // The field's own raw value (space-padded or not); trailing spaces
     // are trimmed internally before front-compression, matching VFP's
     // own observed character-key storage convention.
